apps/ai-service/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9
httpx>=0.27.0
Pillow>=10.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
boto3>=1.34.0
numpy>=1.24.0

# AI / GPU – optional: pip install -r requirements-gpu.txt for local SDXL (CUDA, ~10GB+ VRAM)
# Without GPU deps, SDXLGenerationService uses stubs (placeholder images).
# torch>=2.0.0
# diffusers>=0.30.0
# transformers>=4.40.0
# accelerate>=0.33.0
# safetensors>=0.4.0
